log服务器，在pc上显示设备日志

+++ kline/src/main/java/com/android/ganyue/kline/DayInfo.java
@@ -9,9 +9,9 @@ public class DayInfo {
     public float high;  // 08 ~ 11 字节：最高价
     public float low;   // 12 ~ 15 字节：最低价
     public float close; // 16 ~ 19 字节：收盘价
-    public float volA;    // 20 ~ 23 字节：成交额（元）
-    public int volV;    // 24 ~ 27 字节：成交量（手）
-    public int extra;   // 28 ~ 31 字节：保留，一般都是0
+    public double volA;    // 20 ~ 23 字节：成交额（元）
+    public long volV;    // 24 ~ 27 字节：成交量（手）
+    public float extra;   // 28 ~ 31 字节：保留，一般都是0
 
     public float rate;  // 涨幅 （当天收盘价 - 前天收盘价）/ 前天收盘价
     public int preDate; // 上个交易日的日期
